requirements.txt
openai>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
chromadb>=0.4.0
python-dotenv>=1.0.0
japanize-matplotlib>=1.1.3

